+++ src/NUnitFramework/testdata/FixtureDependsOnAttributeFixture.cs
@@ -27,19 +27,6 @@ namespace NUnit.TestData
 
     public class FixtureDependencyBase
     {
-        [OneTimeSetUp]
-        public void SetUp()
-        {
-            var testName = TestContext.CurrentContext.Test.Name;
-            FixtureDependencyEvents.Record($"{testName}.OneTimeSetUp");
-        }
-
-        [OneTimeTearDown]
-        public void TearDown()
-        {
-            var testName = TestContext.CurrentContext.Test.Name;
-            FixtureDependencyEvents.Record($"{testName}.OneTimeTearDown");
-        }
     }
 
     #region Basic Functionality
@@ -65,7 +52,7 @@ namespace NUnit.TestData
     }
 
     [TestFixture]
-    public class FixtureDependencyBeforeFailing : FixtureDependencyBase
+    public class FixtureDependencyBeforeFailing
     {
         [Test]
         public void BeforeFailingTest()
@@ -76,7 +63,7 @@ namespace NUnit.TestData
 
     [TestFixture]
     [DependsOn(typeof(FixtureDependencyBeforeFailing))]
-    public class FixtureDependencyAfterFailing : FixtureDependencyBase
+    public class FixtureDependencyAfterFailing
     {
         [Test]
         public void AfterFailingDependencyTest()
@@ -87,7 +74,7 @@ namespace NUnit.TestData
 
     [TestFixture]
     [DependsOn(typeof(FixtureDependencyBeforeFailing), AllowFailure = true)]
-    public class FixtureDependencyAfterFailingAllowed : FixtureDependencyBase
+    public class FixtureDependencyAfterFailingAllowed
     {
         [Test]
         public void AfterFailingDependencyAllowedTest()
@@ -97,7 +84,7 @@ namespace NUnit.TestData
     }
 
     [TestFixture]
-    public class ForkingDependencyRoot : FixtureDependencyBase
+    public class ForkingDependencyRoot
     {
         [Test]
         public void RootTest()
@@ -108,7 +95,7 @@ namespace NUnit.TestData
 
     [TestFixture]
     [DependsOn(typeof(ForkingDependencyRoot))]
-    public class ForkingDependencyNodeA : FixtureDependencyBase
+    public class ForkingDependencyNodeA
     {
         [Test]
         public void NodeATest()
@@ -119,7 +106,7 @@ namespace NUnit.TestData
 
     [TestFixture]
     [DependsOn(typeof(ForkingDependencyRoot))]
-    public class ForkingDependencyNodeB : FixtureDependencyBase
+    public class ForkingDependencyNodeB
     {
         [Test]
         public void NodeBTest()
@@ -155,7 +142,7 @@ namespace NUnit.TestData
 
     [TestFixture]
     [DependsOn(typeof(FixtureDependencyCycleA))]
-    public class FixtureDependencyCycleReferrer : FixtureDependencyBase
+    public class FixtureDependencyCycleReferrer
     {
         [Test]
         public void C()
@@ -267,294 +254,5 @@ namespace NUnit.TestData
         }
     }
 
-    #endregion
-
-    #region Method Dependencies
-    [TestFixture]
-    public class MethodDependencyOrdered
-    {
-        [Test]
-        [DependsOn(nameof(Before))]
-        public void After()
-        {
-            FixtureDependencyEvents.Record(nameof(After));
-            Assert.That(true, Is.True);
-        }
-
-        [Test]
-        public void Before()
-        {
-            FixtureDependencyEvents.Record(nameof(Before));
-            Assert.That(true, Is.True);
-        }
-    }
-
-    [TestFixture]
-    public class MethodDependencyFailing
-    {
-        [Test]
-        public void BeforeFailing()
-        {
-            FixtureDependencyEvents.Record(nameof(BeforeFailing));
-            Assert.Fail("Intentional failure for method dependency behavior testing");
-        }
-
-        [Test]
-        [DependsOn(nameof(BeforeFailing))]
-        public void AfterFailing()
-        {
-            FixtureDependencyEvents.Record(nameof(AfterFailing));
-            Assert.That(true, Is.True);
-        }
-    }
-
-    [TestFixture]
-    public class MethodDependencyFork
-    {
-        [Test]
-        public void Root()
-        {
-            FixtureDependencyEvents.Record(nameof(Root));
-            Assert.That(true, Is.True);
-        }
-
-        [Test]
-        [DependsOn(nameof(Root))]
-        public void NodeA()
-        {
-            FixtureDependencyEvents.Record(nameof(NodeA));
-            Assert.That(true, Is.True);
-        }
-
-        [Test]
-        [DependsOn(nameof(Root))]
-        public void NodeB()
-        {
-            FixtureDependencyEvents.Record(nameof(NodeB));
-            Assert.That(true, Is.True);
-        }
-    }
-
-    [TestFixture]
-    public class MethodDependencyFailingAllowed
-    {
-        [Test]
-        public void BeforeFailing()
-        {
-            FixtureDependencyEvents.Record(nameof(BeforeFailing));
-            Assert.Fail("Intentional failure for method dependency behavior testing");
-        }
-
-        [Test]
-        [DependsOn(nameof(BeforeFailing), AllowFailure = true)]
-        public void AfterFailingAllowed()
-        {
-            FixtureDependencyEvents.Record(nameof(AfterFailingAllowed));
-            Assert.That(true, Is.True);
-        }
-    }
-
-    [TestFixture]
-    public class MethodDependencyMissing
-    {
-        [Test]
-        [DependsOn(nameof(NotATestMethod))]
-        public void DependsOnMissingMethod()
-        {
-            FixtureDependencyEvents.Record(nameof(DependsOnMissingMethod));
-            Assert.That(true, Is.True);
-        }
-
-        private void NotATestMethod()
-        {
-        }
-    }
-
-    [TestFixture]
-    public class MethodDependencyCycle
-    {
-        [Test]
-        [DependsOn(nameof(B))]
-        public void A()
-        {
-            FixtureDependencyEvents.Record(nameof(A));
-            Assert.That(true, Is.True);
-        }
-
-        [Test]
-        [DependsOn(nameof(A))]
-        public void B()
-        {
-            FixtureDependencyEvents.Record(nameof(B));
-            Assert.That(true, Is.True);
-        }
-
-        [Test]
-        [DependsOn(nameof(A))]
-        public void Referrer()
-        {
-            FixtureDependencyEvents.Record(nameof(Referrer));
-            Assert.That(true, Is.True);
-        }
-    }
-
-    [TestFixture]
-    public class MethodDependencySelfReferential
-    {
-        [Test]
-        [DependsOn(nameof(Self))]
-        public void Self()
-        {
-            FixtureDependencyEvents.Record(nameof(Self));
-            Assert.That(true, Is.True);
-        }
-    }
-
-    [TestFixture]
-    public class MethodDependencyOrderTarget
-    {
-        [Test]
-        [Order(1)]
-        public void Target()
-        {
-            FixtureDependencyEvents.Record(nameof(Target));
-            Assert.That(true, Is.True);
-        }
-
-        [Test]
-        [DependsOn(nameof(Target))]
-        public void Referrer()
-        {
-            FixtureDependencyEvents.Record(nameof(Referrer));
-            Assert.That(true, Is.True);
-        }
-    }
-
-    [TestFixture]
-    public class MethodDependencyOrderReferrer
-    {
-        [Test]
-        public void Before()
-        {
-            FixtureDependencyEvents.Record(nameof(Before));
-            Assert.That(true, Is.True);
-        }
-
-        [Test]
-        [DependsOn(nameof(Before))]
-        [Order(1)]
-        public void After()
-        {
-            FixtureDependencyEvents.Record(nameof(After));
-            Assert.That(true, Is.True);
-        }
-    }
-
-    [TestFixture]
-    public class MethodDependencyOrderIndependent
-    {
-        [Test]
-        public void Before()
-        {
-            FixtureDependencyEvents.Record(nameof(Before));
-            Assert.That(true, Is.True);
-        }
-
-        [Test]
-        [DependsOn(nameof(Before))]
-        public void After()
-        {
-            FixtureDependencyEvents.Record(nameof(After));
-            Assert.That(true, Is.True);
-        }
-
-        [Test]
-        [Order(1)]
-        public void OrderedIndependent()
-        {
-            FixtureDependencyEvents.Record(nameof(OrderedIndependent));
-            Assert.That(true, Is.True);
-        }
-    }
-
-    [TestFixture]
-    public class MethodDependencyParallel
-    {
-        [Test]
-        [Parallelizable(ParallelScope.Self)]
-        public void BeforeSlow()
-        {
-            Thread.Sleep(150);
-            FixtureDependencyEvents.Record(nameof(BeforeSlow));
-            Assert.That(true, Is.True);
-        }
-
-        [Test]
-        [Parallelizable(ParallelScope.Self)]
-        [DependsOn(nameof(BeforeSlow))]
-        public void AfterParallelDependency()
-        {
-            FixtureDependencyEvents.Record(nameof(AfterParallelDependency));
-            Assert.That(true, Is.True);
-        }
-
-        [Test]
-        public void IndependentTest()
-        {
-            FixtureDependencyEvents.Record(nameof(IndependentTest));
-            Assert.That(true, Is.True);
-        }
-    }
-
-    [TestFixture]
-    [DependsOn(nameof(FixtureMethod))]
-    public class MethodDependencyInvalidStringOnFixture
-    {
-        [Test]
-        public void FixtureMethod()
-        {
-            Assert.That(true, Is.True);
-        }
-    }
-
-    [TestFixture]
-    public class MethodDependencyInvalidTypeOnMethod
-    {
-        [Test]
-        public void DependencyTarget()
-        {
-            Assert.That(true, Is.True);
-        }
-
-        [Test]
-        [DependsOn(typeof(MethodDependencyInvalidTypeOnMethod))]
-        public void DependantMethod()
-        {
-            Assert.That(true, Is.True);
-        }
-    }
-
-    [TestFixture]
-    public class MethodDependencyBase
-    {
-        [Test]
-        public virtual void B()
-        {
-            FixtureDependencyEvents.Record(nameof(B));
-            Assert.That(true, Is.True);
-        }
-    }
-
-    [TestFixture]
-    public class MethodDependencySelfReferentialBase : MethodDependencyBase
-    {
-        [Test]
-        [DependsOn(nameof(B))]
-        public override void B()
-        {
-            FixtureDependencyEvents.Record(nameof(B));
-            Assert.That(true, Is.True);
-        }
-    }
     #endregion
 }
